Refactor Packages in storage test
Make the packages more organised.
