+++ src/main/java/seedu/address/storage/connection/JsonAdaptedPersonMeetingConnection.java
@@ -6,7 +6,6 @@ import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.parser.DateTimeUtil;
 import seedu.address.model.connection.PersonMeetingConnection;
 import seedu.address.model.meeting.*;
-import seedu.address.model.person.AddressBook;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.PersonName;
 import seedu.address.model.person.ReadOnlyAddressBook;
@@ -23,7 +22,7 @@ public class JsonAdaptedPersonMeetingConnection {
     private final String meetingName;
 
     /**
-     * Constructs a {@code JsonAdoptedMeeting} with the given meeting details.
+     * Constructs a {@code JsonAdoptedPersonMeetingConnection} with the given associated person and meeting
      */
 
     @JsonCreator
@@ -37,9 +36,10 @@ public class JsonAdaptedPersonMeetingConnection {
 
 
     /**
-     * Converts a given {@code Meeting} into this class for Jackson use.
+     * Constructs a json serializable association class for the meeting to person.
      */
-    public JsonAdaptedPersonMeetingConnection(Meeting meetingSource, Person personSource) {
+    public JsonAdaptedPersonMeetingConnection(Person personSource, Meeting meetingSource) {
+        assert personSource != null && meetingSource != null;
         personName = personSource.getName().fullName;
         startDateTime = DateTimeUtil.formatDateTime(meetingSource.getStart().value);
         meetingName = meetingSource.getName().fullName;
@@ -53,6 +53,7 @@ public class JsonAdaptedPersonMeetingConnection {
 
     public PersonMeetingConnection toModelType(ReadOnlyAddressBook addressBook, ReadOnlyMeetingBook meetingBook,
                                                PersonMeetingConnection connection) throws IllegalValueException {
+        assert addressBook != null && meetingBook != null && connection != null;
         Person person = addressBook.getPersonByName(new PersonName(personName));
         Meeting meeting = meetingBook.getMeetingByNameAndStartTime(new MeetingName(meetingName), new DateTime(startDateTime));
         if (person == null) {
