+++ src/test/java/seedu/address/storage/meetings/JsonMeetingBookStorageTest.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.meetings;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.address.testutil.Assert.assertThrows;
